requirements.txt
# IA e Integrações
groq
python-dotenv

# Utilitários
requests